+++ liveavatar/models/wan/wan_2_2/modules/s2v/motioner.py
@@ -31,10 +31,10 @@ def sinusoidal_embedding_1d(dim, position):
 def rope_params(max_seq_len, dim, theta=10000):
     assert dim % 2 == 0
     freqs = torch.outer(
-        torch.arange(max_seq_len),
+        torch.arange(max_seq_len, dtype=torch.float32),
         1.0 / torch.pow(theta,
-                        torch.arange(0, dim, 2).to(torch.float64).div(dim)))
-    freqs = torch.polar(torch.ones_like(freqs), freqs)
+                        torch.arange(0, dim, 2).to(torch.float32).div(dim)))
+    freqs = torch.polar(torch.ones_like(freqs), freqs).to(torch.complex64)
     return freqs
 
 
@@ -104,7 +104,7 @@ def rope_apply(x, grid_sizes, freqs, start=None):
                 # precompute multipliers
                 x_i = torch.view_as_complex(
                     x[i, seq_bucket[-1]:seq_bucket[-1] + seq_len].to(
-                        torch.float64).reshape(seq_len, n, -1, 2))
+                        torch.float32).reshape(seq_len, n, -1, 2))
                 x_i = torch.view_as_real(x_i * freqs_i).flatten(2)
                 output[i, seq_bucket[-1]:seq_bucket[-1] + seq_len] = x_i
         seq_bucket.append(seq_bucket[-1] + seq_len)
